# Completely dynamic JS hints