--- a/src/ui/zabapgit_css_theme_default.w3mi.data.css
+++ b/src/ui/zabapgit_css_theme_default.w3mi.data.css
@@ -16,6 +16,7 @@
   --theme-greyscale-medium: #b3b3b3;
   --theme-greyscale-light: #ccc;
   --theme-greyscale-lighter: lightgrey;
+  --theme-linkhint-background: lightgreen;
 }
 
 /* GLOBALS */
@@ -335,6 +336,8 @@ div.info-panel .update {
 
 /* Tooltips text */
 .link-hint { color: var(--theme-primary-font-color); }
+.link-hint { background-color: var(--theme-linkhint-background) }
+.link-hint::after { border-top-color: var(--theme-linkhint-background) }
 
 /* HOTKEYS */
 ul.hotkeys span.key-id {