to #2811

@christianguenter2 

draft for the moment, might be some polishing needed but works and ready for review

- hints are completely generated in JS, so no auxilliary `<span class="tooltiptext hidden">`
- some terminology fixes (tooltip vs hint) and some css
- I changed the JS code quite heavily ... sorry ... appeared to be an interesting subject, carried away :) ... hopefully for good.
- `ZCL_ABAPGIT_HTML ` - just removed the tooltip span, others are se80 reorderings
- same for ZCL_ABAPGIT_GUI_PAGE - just a rename in JS function
- known issue: a dropdown displayed with hint cannot be hidden ... actually no change from existing logic ... not sure if possible to fix cleanly ... maybe with timeout
- known issue: I added backspace for reverting hint steps back ... but works only on the first page - after IE perceives it as back button :( not sure how to fix

Doubt: why setting color in settings ? Imho it should depend on theme directly. Didn't change but imho not quite consistent.