@sbcgua 2019-07-30
P.S. I think JS file should be splitted in parts ... maintenance is quite inconvenient by now