@larshp 2019-07-31
what is hints?