@sbcgua 2019-07-31
The keyboard navigation feature Christian mentioned in #2775 
![image](https://user-images.githubusercontent.com/15635498/62192441-4d8f5100-b37e-11e9-9be7-bf2c254f33a6.png)