@larshp 2019-07-31
any impacts on performance with these changes?