@sbcgua 2019-07-31
No sure how to measure. In theory it should be faster because of pointers caching. But not sure how to measure it properly, especially in IE. I'm open for suggestions :)