@christianguenter2 2019-07-31
> Doubt: why setting color in settings ? Imho it should depend on theme directly. Didn't change but imho not quite consistent.

Agree. I guess there's the option to set the color in the chrome vimium plugin. I thought it would be cool to have the same in abapGit. But to be honest, I didn't change the standard color since then. And we hadn't themes when I implemented this ;-)