@christianguenter2 2019-07-31
> I changed the JS code quite heavily ... sorry ... appeared to be an interesting subject, carried away :) ... 

Sure, no problem at all :-) I can learn a lot.