@christianguenter2 2019-07-31
> known issue: a dropdown displayed with hint cannot be hidden ... actually no change from existing logic ... not sure if possible to fix cleanly ... maybe with timeout

I can be hidden with the hint. That's ok for me.