@christianguenter2 2019-07-31
> any impacts on performance with these changes?
Didn't notice any negative impact during my tests.