@FreHu 2019-07-31
@sbcgua
>No sure how to measure. In theory it should be faster because of pointers caching and less dynamic dom changes. But not sure how to measure it properly, especially in IE. I'm open for suggestions :)

You can profile using iechooser, although you need to export it and then import it in a normal IE window.