@larshp 2019-07-31
looks like the JS is only run when hotkeys are enabled, good,
![image](https://user-images.githubusercontent.com/5888506/62212616-63fad400-b3a1-11e9-9afa-13770667a7b1.png)