@sbcgua 2019-07-31
@christianguenter2 

- color: ok, if you don't mind I'll remove it then :) 
- what do you think about backspace feature ? did you try ? I thought it would be nice to undo typings ... but since it works only on the first page (most frequently the tutorial ...) it does not make sense :( I'm considering removing it
- "dropdown can be hidden with the hint" - hmm, you mean remembering the last triggered dropdown and clearing it on hotkey press ? Interesting idea. Thought can be a bit tricky - I'll have a look. P.S. sadly `li` does not support `:focus` state - it would solve the issue cleanly, just click anywhere and focus changes together with CSS