@sbcgua 2019-07-31
@FreHu @larshp did some performance tests, with good old `console.time`

Scenario: init (obviously), show all hints, hide all hints, show all again, type 146 for filtering (each time less items to filter), total number of hints 3466. last 2000 hints were one above another ... made a separate test to check if block elements would be different but the difference seems to be minor.

Stages below named differently for new and old logic, but follow the same order - it has the same meaning

IE
```
New logic
deployHintContainers: 421ms
displayHints true:    489,9ms
displayHints false:   152,5ms
displayHints true:    441,4ms
filterHints 1:        283,8ms
filterHints 4:        59ms
filterHints 6:        44,2ms

Old logic
querySelectorAll:       2ms
toggleAllTooltips true: 471,2ms
toggleAllTooltips false: 401,8ms
toggleAllTooltips true: 399,2ms
filterTooltips 1:       326,7ms
filterTooltips 4:       77,6ms
filterTooltips 6:       44,8ms
```

Initialization is obviously taking more time in the new logic as it creates elements. Though half a second per 3,5k hints ... seems to be OK for me. Displaying, surprisingly for me, is a bit longer then in old one. Seems like because internal structure of the hint in the new logic is more complex (2 independent spans) - made a quick test for this. Hiding is faster in new logic. And also filtering (probably because of those 2 independent spans - back side of the coin - that was why I actually made them). But in general - same figures imho.

Now to extend the study, let's run it in Chrome

```
New logic
deployHintContainers: 17.22ms
displayHints true:    19.35ms
displayHints false:    9.38ms
displayHints true:    18.66ms
filterHints 1:        22.15ms
filterHints 4:        10.61ms
filterHints 6:         8.61ms

Old logic

querySelectorAll:         0.21ms
toggleAllTooltips true:  54.86ms
toggleAllTooltips false: 68.09ms
toggleAllTooltips true:  35.89ms
filterTooltips 1:       127.24ms
filterTooltips 4:        72.48ms
filterTooltips 6:       105.40ms
```
Yeah ... was is the same PC ? :)