@christianguenter2 2019-08-01
> * color: ok, if you don't mind I'll remove it then :)
okay

> * what do you think about backspace feature ? did you try ? I thought it would be nice to undo typings ... but since it works only on the first page (most frequently the tutorial ...) it does not make sense :( I'm considering removing it
Yeah, if I misstype something I press the activation hotkey twice and start over again.

> * "dropdown can be hidden with the hint" - hmm, you mean remembering the last triggered dropdown and clearing it on hotkey press ? Interesting idea. Thought can be a bit tricky - I'll have a look. P.S. sadly `li` does not support `:focus` state - it would solve the issue cleanly, just click anywhere and focus changes together with CSS

Note quite sure whether we are talking about the same thing. I mean this:
![abapGit5](https://user-images.githubusercontent.com/17437789/62325679-8299c180-b4ac-11e9-80d7-5427da1a3eb5.gif)