@sbcgua 2019-08-01
OK, committed the discussed changes.

> Dropdown logic

Ah, OK, I see what you mean. This should also work :)
But I also added some minor improvements in the last commits. 
- if you force-close the hints (second "f"), and a dropdown was displayed - it is closed automatically. 
- Also if you open a dropdown, and show hints again (like in your gif), and mistype the command, then all hints are closed (it was so before) and also opened dropdown closes too.

I release the PR from draft, imho it is complete. But I'll be away for a week, so no bugcare :)) Maybe wait a bit with the merge if you prefer :))