@christianguenter2 2019-08-02
Small bug. When a page is newly rendered, hint on a dropdown doesn't work the first time.
![abapGit6](https://user-images.githubusercontent.com/17437789/62359892-0210aa80-b518-11e9-95dd-d534cf29dcf7.gif)

I try to fix it.